refactor some of injection script into utils file
